(web) 🗃️ ftree: delete utils folder
