--- apps/web/src/utils/formatText.ts
@@ -1,10 +1,0 @@
-import { REGEX } from '@/constants'
-
-export function formatText(text: string) {
-  const { betweenTwoAsterisks, betweenFourAsterisks } = REGEX
-  const strongComponent = '<span class="strong">$1</span>'
-
-  return text
-    .replace(betweenTwoAsterisks, strongComponent)
-    .replace(betweenFourAsterisks, strongComponent)
-}
