--- apps/web/src/utils/getAppBaseUrl.ts
@@ -1,7 +1,0 @@
-import { ENV } from '@/constants'
-
-export function getAppBaseUrl() {
-  const host = ENV.appHost
-  const url = host.includes('http') ? host : `https://${host}`
-  return url
-}
