--- apps/web/src/utils/index.ts
@@ -1,3 +1,0 @@
-export { formatText } from './formatText'
-export { getAppBaseUrl } from './getAppBaseUrl'
-export { getTemplateContent } from './getTemplateContent'
